rpi/requirements.txt
pyserial
requests
python-dotenv
paho-mqtt